+++ samples/Samples.DirectMessages/Program.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Diagnostics;
 using System.Text;
 
 using idunno.AtProto;
@@ -43,7 +42,7 @@ public class Program
         // proxyUri = new Uri("http://localhost:8888");
 
         // Change the log level in the ConfigureConsoleLogging() to enable logging
-        using (ILoggerFactory? loggerFactory = Helpers.ConfigureConsoleLogging(LogLevel.Debug))
+        using (ILoggerFactory? loggerFactory = Helpers.ConfigureConsoleLogging(LogLevel.Error))
 
         // Create a new BlueSkyAgent
         using (var agent = new BlueskyAgent(
@@ -91,17 +90,9 @@ public class Program
                 }
             }
 
-            var startConversationResult = await agent.GetConversationForMembers(["did:plc:hfgp6pj3akhqxntgqwramlbg"], cancellationToken: cancellationToken);
-            if (startConversationResult.Succeeded)
-            {
-                var post = await agent.GetPostRecord("at://did:plc:hfgp6pj3akhqxntgqwramlbg/app.bsky.feed.post/3lqxyqocwx22m", cancellationToken: cancellationToken);
-
-                var sendMessageResult = await agent.SendMessage(
-                    startConversationResult.Result.Id,
-                    "Embedded post test",
-                    embeddedPost: post.Result!.StrongReference,
-                    cancellationToken: cancellationToken);
-            }
+            var getUnreadCountsResult = await agent.GetUnreadCounts(includeGroupChats: true, cancellationToken: cancellationToken);
+            getUnreadCountsResult.EnsureSucceeded();
+            Console.WriteLine($"Unread Accepted Chats: {getUnreadCountsResult.Result.UnreadAcceptedConversations} | Unread Requested Chats: {getUnreadCountsResult.Result.UnreadRequestedConversations}\n");
 
             var listConversations = await agent.ListConversations(cancellationToken: cancellationToken);
 
@@ -116,20 +107,25 @@ public class Program
                             break;
                         }
 
-                        StringBuilder conversationMembers = new();
-                        foreach (ProfileViewBasic member in conversation.Members)
-                        {
-                            conversationMembers.Append(member.ToString());
-                            conversationMembers.Append(" & ");
-                        }
-                        conversationMembers.Length -= 3;
-
                         if (conversation.Status == ConversationStatus.Requested)
                         {
                             Console.Write("\u001b[1mRequested ");
                         }
 
-                        Console.Write($"Conversation #{conversation.Id} between {conversationMembers}");
+
+                        if (conversation.Kind is GroupConversation groupConversation)
+                        {
+                            Console.WriteLine($"Group Chat: {groupConversation.Name}");
+                            Console.WriteLine($"  Member Count: {groupConversation.MemberCount}");
+                        }
+                        else if (conversation.Kind is DirectConversation directConversation)
+                        {
+                            Console.WriteLine($"Direct Conversation");
+                        }
+
+                        string conversationMembers = string.Join(", ", conversation.Members.Select(m => m.Handle));
+
+                        Console.Write($"  Conversation #{conversation.Id} between {conversationMembers}");
 
                         if (conversation.UnreadCount != 0)
                         {
@@ -210,6 +206,8 @@ public class Program
                                              !string.IsNullOrEmpty(getMessages.Result.Cursor));
                                 }
 
+                                Console.WriteLine();
+
                                 await agent.UpdateRead(conversation.Id, cancellationToken: cancellationToken);
                             }
                         }
@@ -219,13 +217,10 @@ public class Program
                     {
                         listConversations = await agent.ListConversations(cursor: listConversations.Result.Cursor, cancellationToken: cancellationToken);
                     }
-
                 } while (!cancellationToken.IsCancellationRequested &&
                          listConversations.Succeeded &&
                          !string.IsNullOrEmpty(listConversations.Result.Cursor));
             }
-
-            Debugger.Break();
         }
     }
 }