+++ src/idunno.Bluesky/Chat/BlueskyServer.cs
@@ -37,9 +37,6 @@ public static partial class BlueskyServer
     // https://github.com/bluesky-social/atproto/blob/main/lexicons/chat/bsky/convo/getConvoMembers.json
     private const string GetConvoMembersEndpoint = "/xrpc/chat.bsky.convo.getConvoMembers";
 
-    // https://github.com/bluesky-social/atproto/blob/main/lexicons/chat/bsky/convo/getUnreadCounts.json
-    private const string GetUnreadCountsEndpoint = "/xrpc/chat.bsky.convo.getUnreadCounts";
-
     // https://docs.bsky.app/docs/api/chat-bsky-convo-list-convos
     private const string ListConvosEndpoint = "/xrpc/chat.bsky.convo.listConvos";
 
@@ -732,64 +729,6 @@ public static partial class BlueskyServer
         else
         {
             return new AtProtoHttpResult<Messages>(
-                null,
-                response.StatusCode,
-                response.HttpResponseHeaders,
-                response.AtErrorDetail,
-                response.RateLimit);
-        }
-    }
-
-    /// <summary>
-    /// Gets the count of unread conversations for the authenticated user.
-    /// </summary>
-    /// <param name="service">The <see cref="Uri"/> of the service to leave the conversation on.</param>
-    /// <param name="accessCredentials">The <see cref="AccessCredentials"/> to use when accessing the <paramref name="service"/>.</param>
-    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
-    /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
-    /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">
-    /// Thrown when <paramref name="accessCredentials"/>, <paramref name="service"/> or <paramref name="httpClient"/> is <see langword="null"/>.
-    /// </exception>
-    [UnconditionalSuppressMessage(
-        "Trimming",
-        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
-    [UnconditionalSuppressMessage("AOT",
-        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
-    public static async Task<AtProtoHttpResult<UnreadConversationCounts>> GetUnreadCounts(
-        Uri service,
-        AccessCredentials accessCredentials,
-        HttpClient httpClient,
-        Action<AtProtoCredential>? onCredentialsUpdated = null,
-        ILoggerFactory? loggerFactory = default,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(service);
-        ArgumentNullException.ThrowIfNull(accessCredentials);
-        ArgumentNullException.ThrowIfNull(httpClient);
-
-        BlueskyHttpClient<UnreadConversationCounts> client = new(ChatProxy, loggerFactory);
-
-        AtProtoHttpResult<UnreadConversationCounts> response = await client.Get(
-            service,
-            $"{GetUnreadCountsEndpoint}",
-            credentials: accessCredentials,
-            httpClient: httpClient,
-            jsonSerializerOptions: BlueskyJsonSerializerOptions,
-            onCredentialsUpdated: onCredentialsUpdated,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-
-        if (response.Succeeded)
-        {
-            return response;
-        }
-        else
-        {
-            return new AtProtoHttpResult<UnreadConversationCounts>(
                 null,
                 response.StatusCode,
                 response.HttpResponseHeaders,